optimized the code for future!
